requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
anyio=4.3.0=pyhd8ed1ab_0
appnope=0.1.4=pyhd8ed1ab_0
argon2-cffi=23.1.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py312h02f2b3b_4
arrow=1.3.0=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
async-lru=2.0.4=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
babel=2.14.0=pyhd8ed1ab_0
beautifulsoup4=4.12.3=pyha770c72_0
bleach=6.1.0=pyhd8ed1ab_0
brotli=1.1.0=hb547adb_1
brotli-bin=1.1.0=hb547adb_1
brotli-python=1.1.0=py312h9f69965_1
bzip2=1.0.8=h93a5062_5
ca-certificates=2024.2.2=hf0a4a13_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.16.0=py312h8e38eb3_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.1=py312h0fef576_0
cycler=0.12.1=pyhd8ed1ab_0
debugpy=1.8.1=py312h20a0b95_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
ffmpeg=4.2.2=h04105a8_0
filelock=3.14.0=pyhd8ed1ab_0
fonttools=4.51.0=py312he37b823_0
fqdn=1.5.1=pyhd8ed1ab_0
freetype=2.12.1=hadb7bae_2
fsspec=2024.5.0=pyhff2d567_0
gettext=0.22.5=h8fbad5d_2
gettext-tools=0.22.5=h8fbad5d_2
gmp=6.3.0=hebf3989_1
gmpy2=2.1.5=py312hfa9fade_1
gnutls=3.6.15=h887c41c_0
h11=0.14.0=pyhd8ed1ab_0
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
httpcore=1.0.5=pyhd8ed1ab_0
httpx=0.27.0=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_0
idna=3.7=pyhd8ed1ab_0
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
importlib_resources=6.4.0=pyhd8ed1ab_0
ipykernel=6.29.3=pyh3cd1d5f_0
ipython=8.24.0=pyh707e725_0
ipywidgets=8.1.2=pyhd8ed1ab_1
isoduration=20.11.0=pyhd8ed1ab_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pyhd8ed1ab_0
joblib=1.4.2=pyhd8ed1ab_0
json5=0.9.25=pyhd8ed1ab_0
jsonpointer=2.4=py312h81bd7bf_3
jsonschema=4.22.0=pyhd8ed1ab_0
jsonschema-specifications=2023.12.1=pyhd8ed1ab_0
jsonschema-with-format-nongpl=4.22.0=pyhd8ed1ab_0
jupyter=1.0.0=pyhd8ed1ab_10
jupyter-lsp=2.2.5=pyhd8ed1ab_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_console=6.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=py312h81bd7bf_0
jupyter_events=0.10.0=pyhd8ed1ab_0
jupyter_server=2.14.0=pyhd8ed1ab_0
jupyter_server_terminals=0.5.3=pyhd8ed1ab_0
jupyterlab=4.2.1=pyhd8ed1ab_0
jupyterlab_pygments=0.3.0=pyhd8ed1ab_1
jupyterlab_server=2.27.2=pyhd8ed1ab_0
jupyterlab_widgets=3.0.10=pyhd8ed1ab_0
kiwisolver=1.4.5=py312h389731b_1
krb5=1.21.2=h92f50d5_0
lame=3.100=h1a8c8d9_1003
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libabseil=20240116.2=cxx17_hebf3989_0
libasprintf=0.22.5=h8fbad5d_2
libasprintf-devel=0.22.5=h8fbad5d_2
libblas=3.9.0=22_osxarm64_openblas
libbrotlicommon=1.1.0=hb547adb_1
libbrotlidec=1.1.0=hb547adb_1
libbrotlienc=1.1.0=hb547adb_1
libcblas=3.9.0=22_osxarm64_openblas
libcxx=17.0.6=h5f092b4_0
libdeflate=1.20=h93a5062_0
libedit=3.1.20191231=hc8eb9b7_2
libexpat=2.6.2=hebf3989_0
libffi=3.4.2=h3422bc3_5
libgettextpo=0.22.5=h8fbad5d_2
libgettextpo-devel=0.22.5=h8fbad5d_2
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libiconv=1.17=h0d3ecfb_2
libidn2=2.3.7=h93a5062_0
libintl=0.22.5=h8fbad5d_2
libintl-devel=0.22.5=h8fbad5d_2
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=22_osxarm64_openblas
libopenblas=0.3.27=openmp_h6c19121_0
libopus=1.3.1=h27ca646_1
libpng=1.6.43=h091b4b1_0
libprotobuf=4.25.3=hbfab5d5_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.45.3=h091b4b1_0
libtasn1=4.19.0=h1a8c8d9_0
libtiff=4.6.0=h07db509_3
libtorch=2.3.0=cpu_generic_hed86ac2_0
libunistring=0.9.10=h3422bc3_0
libuv=1.48.0=h93a5062_0
libvpx=1.10.0=hc377ac9_0
libwebp-base=1.4.0=h93a5062_0
libxcb=1.15=hf346824_0
libzlib=1.2.13=h53f4e23_5
llvm-openmp=18.1.5=hde57baf_0
markupsafe=2.1.5=py312he37b823_0
matplotlib=3.8.4=py312h1f38498_2
matplotlib-base=3.8.4=py312h4479663_2
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mistune=3.0.2=pyhd8ed1ab_0
mpc=1.3.1=h91ba8db_0
mpfr=4.2.1=h41d338b_1
mpmath=1.3.0=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.10.0=pyhd8ed1ab_0
nbconvert=7.16.4=hd8ed1ab_0
nbconvert-core=7.16.4=pyhd8ed1ab_0
nbconvert-pandoc=7.16.4=hd8ed1ab_0
nbformat=5.10.4=pyhd8ed1ab_0
ncurses=6.5=hb89a1cb_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.7.3=h84b5d62_1
networkx=3.3=pyhd8ed1ab_1
nomkl=1.0=h5ca1d4c_0
notebook=7.2.0=pyhd8ed1ab_0
notebook-shim=0.2.4=pyhd8ed1ab_0
numpy=1.26.4=py312h8442bc7_0
openh264=1.8.0=h98b2900_0
openjpeg=2.5.2=h9f1df11_0
openssl=3.3.0=hfb2fe0b_3
overrides=7.7.0=pyhd8ed1ab_0
packaging=24.0=pyhd8ed1ab_0
pandas=2.2.2=py312h8ae5369_1
pandoc=3.2=hce30654_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_0
patsy=0.5.6=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=py312h8a801b1_0
pip=24.0=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=4.2.2=pyhd8ed1ab_0
prometheus_client=0.20.0=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
psutil=5.9.8=py312he37b823_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.22=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_0
pyobjc-core=10.2=py312h9d22092_0
pyobjc-framework-cocoa=10.2=py312h9d22092_0
pyparsing=3.1.2=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.12.3=h4a7b5fc_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-fastjsonschema=2.19.1=pyhd8ed1ab_0
python-graphviz=0.20.3=pypi_0
python-json-logger=2.0.7=pyhd8ed1ab_0
python-tzdata=2024.1=pyhd8ed1ab_0
python_abi=3.12=4_cp312
pytorch=2.3.0=cpu_generic_py312h16b77fb_0
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.1=py312h02f2b3b_1
pyzmq=26.0.3=py312hfa13136_0
qtconsole-base=5.5.2=pyha770c72_0
qtpy=2.4.1=pyhd8ed1ab_0
readline=8.2=h92ec313_1
referencing=0.35.1=pyhd8ed1ab_0
requests=2.32.2=pyhd8ed1ab_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rpds-py=0.18.1=py312h552d48e_0
scikit-learn=1.4.2=py312h1b546db_1
scipy=1.13.1=py312h14ffa8f_0
seaborn=0.13.2=hd8ed1ab_2
seaborn-base=0.13.2=pyhd8ed1ab_2
send2trash=1.8.3=pyh31c8845_0
setuptools=70.0.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h156473d_2
sniffio=1.3.1=pyhd8ed1ab_0
soupsieve=2.5=pyhd8ed1ab_1
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.2=py312hbebd99a_0
sympy=1.12=pypyh9d50eac_103
terminado=0.18.1=pyh31c8845_0
threadpoolctl=3.5.0=pyhc1e730c_0
tinycss2=1.3.0=pyhd8ed1ab_0
tk=8.6.13=h5083fa2_1
tomli=2.0.1=pyhd8ed1ab_0
torchvision=0.18.0=cpu_py312ha63f1fc_0
torchviz=0.0.2=pypi_0
tornado=6.4=py312he37b823_0
tqdm=4.66.4=pyhd8ed1ab_0
traitlets=5.14.3=pyhd8ed1ab_0
types-python-dateutil=2.9.0.20240316=pyhd8ed1ab_0
typing-extensions=4.11.0=hd8ed1ab_0
typing_extensions=4.11.0=pyha770c72_0
typing_utils=0.1.0=pyhd8ed1ab_0
tzdata=2024a=h0c530f3_0
uri-template=1.3.0=pyhd8ed1ab_0
urllib3=2.2.1=pyhd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_0
webcolors=1.13=pyhd8ed1ab_0
webencodings=0.5.1=pyhd8ed1ab_2
websocket-client=1.8.0=pyhd8ed1ab_0
wheel=0.43.0=pyhd8ed1ab_1
widgetsnbextension=4.0.10=pyhd8ed1ab_0
x264=1!152.20180806=h1a28f6b_0
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
zeromq=4.3.5=hcc0f68c_4
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h53f4e23_5
zstd=1.5.6=hb46c0d2_0
